全面升级 优化代码编写逻辑实现,使用代码更简洁可读

+++ future-global/src/main/java/com/suven/framework/http/data/entity/PageResult.java
@@ -170,7 +170,7 @@ public  class PageResult<T> implements IResponseResultPage<T> {
 		return total;
 	}
 
-	public void setTotal(int total) {
+	public void setTotal(long total) {
 		this.total = total;
 	}
